Clone a project repo on demand, if necessary.
Currently the only way a repo is cloned is when a DNProject record
has been saved within DNAdmin. If the clone is removed, an admin
needs to change the CVSPath field in the CMS and save the record
so onBeforeWrite() triggers a clone.

This fixes it so if the repo is setup, but doesn't exist, it will
be cloned on demand, which is the case of viewing naut/project/<myproject>