Fixing permissions on release temp dir prior to update_code cap task.
When Capistrano uses git clone to checkout the repo during update_code
task, there is a bug with Git where it hardcodes mkdir(repo, 0755).
The fix for this is in Git 1.7.11, but Debian still uses 1.7.10, even
on the latest version of the package.

This fix first creates /tmp/{release_name} prior to update_code,
so if there's special permissions (e.g. umask), then it creates the
directory with the correct permissions.