Fixes for running Git commands under an alternate user in deploynaut
Fixes a regression in cc86470e7d2ae2c4c6b651fa8380e89550ba8d92 where
Gitonomy "command" parameter doesn't accept any use of "sudo -u user".
Use Symfony process directly in FetchJob cases.

If another user is cloning a repo, CloneGitRepo will re-create the
local git directory it under the www-data user, which isn't appropriate.
It should all be handled by the other user in this case

git clone already runs mkdir() to create a directory, so mkdir()
shouldn't be needed in CloneGitRepo.

In order for the other user to remove an existing local repo when
CloneGitRepo is called, it needs to be run under that user. To do this,
remove the use of delTree and just "rm -rf" the path directly.