Fixing bug where SyncProjectsAndEnvironments failed on undefined method
DNProject::DNEnvironmentList() returned the wrong type of list, because
filterByCallback changes it to an ArrayList.

The whole thing of overloading the static get() on DNEnvironment and
DNProject to return a specific DNEnvironmentList or DNProjectList
seemed prone to error, so I've removed that and refactored the sync
and remove obsolete methods onto SyncProjectsAndEnvironments. It seems
like they belong here anyway, not on the lists themselves.