NEW: Introduce PackageGenerator, to break out generation of the build package to be separate from the Capistrano deploy.
This first cut includes an implementation, SimplePackageGenerator, that will run a pre-defined build script locally. Other implementations, that execute a docker container and/or pull a build from a CI server, should be able to implemented.

The package generator is linked into the top-level deployment system; it is called by the CapistranoDeploymentBackend, and inserted into it via Injector configuration. Other deployment backends should be able to also use it, using the code change in CapistranoDeploymentBackend as a guide.