Moving preflight validate/fixing of sspak into DNDataArchive
This is in preparation for creating a new DeploymentBackend as a means
to clean things up instead of copying/pasting code around.

The current "preflight" archive checks are not strictly specific to
Capistrano, so the checks should really belong on DNDataArchive and
called from the appropriate backends prior to restoring the archive onto
the target environment.