Sanitise the folder names with DNDataArchive consistently
Framework uses FileNameFilter for sanitising the folder names when
generating them from a value with special characters like spaces, etc,
but the current rules don't work as they replace spaces with underscores
and such. We should rely on FileNameFilter::filter() to do the
sanitising work for us instead of hardcoding those regex rules in
DNDataArchive.