DeploynautLogFile: Sanitise file name to avoid filesystem issues.
This makes sure only alphanumeric, dashes, and dots are allowed to
make their way into the log filenames, and nothing else.

Uses FileNameFilter for the sanitisation job. If the old filename
is still being used, make sure we don't break the existing usage
of showing the old log file, as well as writing into it.