Deploy ui summary (#604)
* add the first version of the deployment diff summary

When users pick a git sha it asks the backend for the deploy plan / diff
and will display that on step 2 of the deployment form.

This change also includes a very simple complete and loading hints on
the step navigation.

* Extract reducers into their own modules

The root reducer were becoming to big to easily understand and change. This change replaces one
reducer with the redux combineReducers() that only permits reducers to manage a certain slice of
the global state.