Fix deployment approval email sending
Due to the fix for 0bb589a8f3f908ba9f91dc56e7c453c127e481ca using emails as
as service doesn't work anymore.

The test has been changed from a pure PHPUnit test case to a SapphireTest and
uses the DB and the TestMailer to check if email was sent.