remove the currentbuild, deployhistory and upcomingdeployment reducers
This change will put all deploy data into a list in the deployment reducer. When some components calls
the getDeploy() action creator it will return it from that list of fetch it from the backend if its missing.

To make sure that this works, we need to update the deployment list on every action to the backend that
updates the state of one deployment (write through cache).