fix direct URL opening of a deployment doesn't always show plan
There is a race condition between fetching deployment and git data. When the
fetching of git is done, the plan reducers zeroes out the plan data. This
causes issues when the deploy data is loaded before the git data.

For the user this shows as the deployment loading fine in the UI and then
the plan and approval sections just disappears.